add base encoder
